updating logic game loop done
