+++ VRshooter/Assets/Prefabs/TeleportTicket.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1f59fef944959c44b8647e026592830e
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
