--- VRshooter/Assets/Scripts/MainMenuManager.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 9a73a7390e5b5da4cae8cc47adcc45a3
