--- VRshooter/Assets/Scripts/MenuManger.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: af4d0b1e06169dc408f2c0691dfe22f0
