+++ VRshooter/ProjectSettings/TagManager.asset
@@ -7,6 +7,7 @@ TagManager:
   - Bullet
   - Enemy
   - Weapon
+  - Ticket
   layers:
   - Default
   - TransparentFX
